requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
docxtpl~=0.16.7